HHH-9475 Cannot mix @MapKey with @Convert - Moved classes to reuse in test case
